+++ cmd/weaver/conf/entrypoint.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+set -e
+
+# Block bad hosts
+cat conf/hosts >> /etc/hosts
+
+exec "$@"
