+++ cmd/weaver/conf/hosts
