+++ cmd/weaver/converter/worker.go
