+++ cmd/weaver/docs/building.md
