+++ cmd/weaver/gcmd/command.go
