+++ cmd/weaver/gcmd/command_test.go
