--- weaver/.dockerignore
@@ -1,14 +1,0 @@
-# Docker
-.dockerfile
-docker-compose.yml
-Dockerfile
-Dockerfile.build
-
-# Docs
-README.md
-docs/
-
-# Others
-**/.DS_STORE
-**/.git*
-**/*.env
