--- weaver/Dockerfile.build
@@ -1,14 +1,0 @@
-FROM golang:1.7-alpine
-MAINTAINER Arachnys <techteam@arachnys.com>
-
-RUN apk add --update git
-
-COPY . /go/src/github.com/arachnys/athenapdf/weaver
-WORKDIR /go/src/github.com/arachnys/athenapdf/weaver
-
-RUN \
-  go get -v -d \
-  && go install -v \
-  && CGO_ENABLED=0 go build -ldflags "-s" -a -installsuffix cgo -o weaver .
-
-CMD ["/bin/sh"]
