--- weaver/docker-compose.yml
@@ -1,6 +1,0 @@
-static:
-  image: arachnysdocker/athenapdf-service
-  ports:
-    - "8080:8080"
-  env_file:
-    - ./conf/sample.env
